pyproject.toml
[tool.poetry]
name = "oracle-of-ammon"
version = "0.1.3"
description = "CLI tool for creating Search APIs."
authors = ["Kyle McLester <kyle.mclester@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/kmcleste/oracle-of-ammon"
repository = "https://github.com/kmcleste/oracle-of-ammon"
keywords = ["fastapi", "haystack", "search", "faq", "typer"]
include = ["LICENSE"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent"
]
packages = [{include = "oracle_of_ammon"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
requests = "^2.28.1"
uvicorn = "^0.20.0"
pandas = "^1.5.2"
fastapi = "^0.88.0"
pynvml = "^11.4.1"
typer = {extras = ["all"], version = "^0.7.0"}
aiofiles = "^22.1.0"
python-multipart = "^0.0.5"
farm-haystack = "^1.12.2"


[tool.poetry.group.dev.dependencies]
black = "22.6.0"
flake8 = "^6.0.0"
pre-commit = "^2.21.0"
pytest = "^7.2.0"
coverage = "^7.0.1"
ipykernel = "^6.19.4"
ipywidgets = "^8.0.4"
httpx = "^0.23.1"
bandit = {extras = ["toml"], version = "^1.7.4"}
locust = "^2.14.0"

[tool.poetry.scripts]
oracle-of-ammon = "oracle_of_ammon.cli.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning"
]

[tool.bandit]
skips = ["B101", "B104", "B501"]
